// models/types.ts
export type Task = {
  id: string;
  body: string;
  status: boolean;
}

// export interface AddFormStatus {
//   addingFormStatus: string;
//   setAddingFormStatus: Function;
//   editInput: string;
//   setEditInput: Function;
//   itemEditingId: string;
// }

// export interface HeaderProps {
//   filter: string;
//   setFilter: Function;
//   setAddingFormStatus: Function;
// }

// export interface PostType {
//   index: number;
//   post: TaskType;
//   setAddingFormStatus: Function;
//   setEditInput: Function;
//   setItemEditingId: Function;
// }

// export type TaskType = {
//   id: string,
//   body: string,
//   status: boolean,
// }

// export interface TaskListProps {
//   tasks: Array<TaskType>;
//   setAddingFormStatus: Function;
//   setEditInput: Function;
//   setItemEditingId: Function;
// }